requirements.txt
streamlit
langchain
langchain_community
python-dotenv
transformers
torch
sentence-transformers
faiss-cpu
pyMuPDF